Seller Profile not complete

+++ app/src/main/java/com/example/unshelf/controller/SellerAuthController.kt
@@ -14,7 +14,7 @@ class SellerAuthController ( // Please modify class diagram and add this class a
         address: String,
         storeName: String,
 
-        rating: Long,
+        rating: Double,
         followers: Int,
         adminVerified: String,
         callback: (Boolean, String?) -> Unit
